added animation to posts

+++ components/post/post.component.tsx
@@ -53,9 +53,11 @@ const Post: React.FC<any> = ({ post }) => {
   }
 
   const deletePost = async () => {
-    if(window.confirm("Are you sure you want to delete this post?")){
+    if (window.confirm("Are you sure you want to delete this post?")) {
       deleteDoc(doc(db, "posts", post.id));
-      deleteObject(ref(storage, `posts/${post.id}/image`))
+      if (post.data().image) {
+        deleteObject(ref(storage, `posts/${post.id}/image`));
+      }
     }
   }
